webapp/requirements.txt
celery>=4.0,<5.0
flask
redis
pymongo
flask_restful
bcrypt
PyJWT
validate_email
uwsgi